+++ frontend/src/pages/main-page/components/create-new-study-button/CreateNewStudyButton.style.tsx
@@ -13,10 +13,17 @@ export const CreateNewStudyButton = styled.button`
 
     width: 70px;
     height: 70px;
+    padding: 8px;
 
     border-radius: 50%;
-    border: 1px solid black;
     background-color: ${theme.colors.primary.base};
-    box-shadow: 4px 4px 0px 0px #cfd8dc;
+
+    &:hover {
+      background-color: ${theme.colors.primary.light};
+    }
+
+    &:active {
+      background-color: ${theme.colors.primary.dark};
+    }
   `}
 `;
