+++ frontend/src/pages/main-page/components/create-new-study-button/CreateNewStudyButton.tsx
@@ -1,6 +1,6 @@
 import { theme } from '@styles/theme';
 
-import * as S from '@main-page/create-new-study-button/CreateNewStudyButton.style';
+import * as S from '@pages/main-page/components/create-new-study-button/CreateNewStudyButton.style';
 
 type CreateNewStudyButtonProps = {
   onClick?: React.MouseEventHandler<HTMLButtonElement>;
@@ -8,7 +8,7 @@ type CreateNewStudyButtonProps = {
 
 const CreateNewStudyButton = ({ onClick }: CreateNewStudyButtonProps) => {
   return (
-    <S.CreateNewStudyButton onClick={onClick}>
+    <S.CreateNewStudyButton onClick={onClick} aria-label="스터디 개설 페이지 이동">
       <svg width="48" height="48" viewBox="0 0 48 48" fill="none" xmlns="http://www.w3.org/2000/svg">
         <path
           d="M5.325 43.5H13.81L44.923 12.387L36.437 3.902L5.325 35.015V43.5Z"
