+++ frontend/src/pages/main-page/components/filter-section/FilterSection.style.tsx
