+++ frontend/src/pages/main-page/components/filter-section/filter-button-list/FilterButtonList.stories.tsx
@@ -1,8 +1,8 @@
 import { Story } from '@storybook/react';
 import { useState } from 'react';
 
-import type { FilterButtonListProps } from '@pages/main-page/filter-section/filter-button-list/FilterButtonList';
-import FilterButtonList from '@pages/main-page/filter-section/filter-button-list/FilterButtonList';
+import type { FilterButtonListProps } from '@pages/main-page/components/filter-section/filter-button-list/FilterButtonList';
+import FilterButtonList from '@pages/main-page/components/filter-section/filter-button-list/FilterButtonList';
 
 export default {
   title: 'Components/FilterButtonList',
