+++ frontend/src/pages/main-page/components/study-card/StudyCard.tsx
